# perf: increase default SWC minimizer passes to reduce bundle size