--- a/packages/rspack-test-tools/tests/configCases/plugins/minify-disable-minify/index.js
+++ b/packages/rspack-test-tools/tests/configCases/plugins/minify-disable-minify/index.js
@@ -4,6 +4,5 @@ const path = require("path");
 it("[minify-disable-minify]: should not minify code", () => {
 	const content = fs.readFileSync(path.resolve(__dirname, "a.js"), "utf-8");
 
-	expect(content).toContain("let a = process.env.a;");
-	expect(content).toContain("console.log(a + process.env.b);");
+	expect(content).toContain("console.log(process.env.a + process.env.b);");
 });