--- a/packages/rspack-test-tools/tests/statsAPICases/ids.js
+++ b/packages/rspack-test-tools/tests/statsAPICases/ids.js
@@ -17,7 +17,7 @@ module.exports = {
 		};
 		expect(stats?.toJson(options)).toMatchSnapshot();
 		expect(stats?.toString(options)).toMatchInlineSnapshot(`
-		asset main.js 207 bytes {909} [emitted] (name: main)
+		asset main.js 206 bytes {909} [emitted] (name: main)
 		chunk {909} (runtime: main) main.js (main) 55 bytes [entry] [rendered]
 		./fixtures/a.js [585] 55 bytes {909} [built] [code generated]
 	`);