… packages/rspack-test-tools/tests/__snapshots__/StatsAPI.test.js.snap
… packages/rspack-test-tools/tests/statsAPICases/child-compiler.js
… packages/rspack-test-tools/tests/statsAPICases/chunks.js
… tests/webpack-test/__snapshots__/StatsTestCases.basictest.js.snap