## Summary

### What is `passes`

The maximum number of times to run compress. In some cases, more than one pass leads to further compressed code. Keep in mind more passes will take more time.

- https://swc.rs/docs/configuration/minification#jscminifycompress
- https://terser.org/docs/options/#compress-options

### Why increase `passes`

Output Size:
- Increasing passes can reduce output size by 0-6%, with optimization results depending on user code
- Optimization effects are similar between 2 and 3 passes, with diminishing returns beyond 3 passes

Build Time:
- For simple cases, increasing passes adds 3-5% to overall build time
- For complex cases, increasing passes adds 10-20% to overall build time

Overall, setting `passes` to `2` by default offers a good balance of build performance and bundle size optimization.

### Test case 1

- Repo: https://github.com/filipsobol/bundle-size-benchmark

- Test config:

```js
output: {
  minify: {
    jsOptions: {
      minimizerOptions: {
        compress: {
          passes: Number(process.env.PASSES) ?? 1,
        },
      },
    },
  },
},
performance: {
  chunkSplit: {
    strategy: "all-in-one",
  },
},
```

- Build time:

```
hyperfine --warmup 2 --runs 5 'PASSES=1 node --run rsbuild' 'PASSES=2 node --run rsbuild' 'PASSES=3 node --run rsbuild' 'PASSES=4 node --run rsbuild'

Benchmark 1: PASSES=1 node --run rsbuild
  Time (mean ± σ):      1.581 s ±  0.018 s    [User: 2.705 s, System: 0.698 s]
  Range (min … max):    1.557 s …  1.604 s    5 runs
 
Benchmark 2: PASSES=2 node --run rsbuild
  Time (mean ± σ):      1.716 s ±  0.030 s    [User: 2.925 s, System: 0.786 s]
  Range (min … max):    1.683 s …  1.762 s    5 runs
 
Benchmark 3: PASSES=3 node --run rsbuild
  Time (mean ± σ):      1.870 s ±  0.067 s    [User: 3.126 s, System: 0.837 s]
  Range (min … max):    1.794 s …  1.962 s    5 runs
 
Benchmark 4: PASSES=4 node --run rsbuild
  Time (mean ± σ):      1.998 s ±  0.042 s    [User: 3.323 s, System: 0.894 s]
  Range (min … max):    1.935 s …  2.045 s    5 runs
 
Summary
  PASSES=1 node --run rsbuild ran
    1.09 ± 0.02 times faster than PASSES=2 node --run rsbuild
    1.18 ± 0.04 times faster than PASSES=3 node --run rsbuild
    1.26 ± 0.03 times faster than PASSES=4 node --run rsbuild
```

- Bundle size:

```
[passes 1]
Total: 2561.6 kB (gzip: 685.3 kB)

[passes 2] 124.6kB less than the previous value
Total: 2437.0 kB (gzip: 650.8 kB)

[passes 3] 29.2kB less than the previous value
Total: 2407.8 kB (gzip: 642.1 kB)

[passes 4] 8.5kB less than the previous value
Total: 2399.3 kB (gzip: 639.3 kB)
```

### Test case 2

- Repo: https://github.com/web-infra-dev/rspack/tree/main/website

- Test config:

```js
  output: {
    minify: {
      jsOptions: {
        minimizerOptions: {
          compress: {
            passes: Number(process.env.PASSES) ?? 1,
          },
        },
      },
    },
  },
```

- Build time:

```
hyperfine --warmup 2 --runs 5 'PASSES=1 node --run build' 'PASSES=2 node --run build' 'PASSES=3 node --run build' 'PASSES=4 node --run build'

Benchmark 1: PASSES=1 node --run build
  Time (mean ± σ):      3.399 s ±  0.165 s    [User: 8.534 s, System: 2.053 s]
  Range (min … max):    3.241 s …  3.654 s    5 runs
 
Benchmark 2: PASSES=2 node --run build
  Time (mean ± σ):      3.539 s ±  0.112 s    [User: 9.080 s, System: 2.199 s]
  Range (min … max):    3.399 s …  3.699 s    5 runs
 
Benchmark 3: PASSES=3 node --run build
  Time (mean ± σ):      3.518 s ±  0.026 s    [User: 9.603 s, System: 2.328 s]
  Range (min … max):    3.491 s …  3.555 s    5 runs
 
Benchmark 4: PASSES=4 node --run build
  Time (mean ± σ):      3.712 s ±  0.114 s    [User: 9.932 s, System: 2.442 s]
  Range (min … max):    3.594 s …  3.885 s    5 runs
 
Summary
  PASSES=1 node --run build ran
    1.04 ± 0.05 times faster than PASSES=3 node --run build
    1.04 ± 0.06 times faster than PASSES=2 node --run build
    1.09 ± 0.06 times faster than PASSES=4 node --run build
```

- Bundle size:

```
[passes 1]
Total: 8950.5 kB (gzip: 2489.8 kB)

[passes 2] 14.3kB less than the previous value
Total: 8936.2 kB (gzip: 2485.5 kB)

[passes 3] 5.7kB less than the previous value
Total: 8930.5 kB (gzip: 2483.5 kB)

[passes 4] 0.4kB less than the previous value
Total: 8930.1 kB (gzip: 2483.3 kB)
```

## Checklist

<!--- Check and mark with an "x" -->

- [ ] Tests updated (or not required).
- [x] Documentation updated (or not required).